pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-local-memory"
version = "0.1.0"
description = "RAM-optimized memory system for coding agents with Obsidian-native markdown storage"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "sqlite-vec>=0.1.6",
    "pysqlite3>=0.6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
memory = "memory.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
